started adding router for blog
