+++ package.json
@@ -11,6 +11,7 @@
     "react-chrono": "^1.22.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "^2.1.3",
     "react-swipeable": "^7.0.0",
     "web-vitals": "^2.1.4"
