+++ src/Components/Blog/BlogCardContainer.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import './BlogCardContainer.css';
-import BlogCard from './BlogCard';
 import Goggins from '../../Assets/MeAndGoggins.JPG';
 import Slider from '../Slider/Slider';
 
@@ -9,19 +8,22 @@ const BlogCardContainer = () => {
     const cards = [
         {
             index: 0,
-            title: "Coming Soon",
+            id: 1,
+            title: "Coming Soon 1",
             category: "Life",
             image: Goggins
         },
         {
             index: 1,
-            title: "Coming Soon",
+            id: 2,
+            title: "Coming Soon 2",
             category: "Life",
             image: Goggins
         },
         {
             index: 2,
-            title: "Coming Soon",
+            id: 3,
+            title: "Coming Soon 3",
             category: "Life",
             image: Goggins
         },
