+++ src/Components/index.js
@@ -1,3 +1,2 @@
-export { default as BlogCard } from './Blog/BlogCard';
 export { default as Hero } from './Hero/Hero';
-export { default as NavBar } from './NavBar/NavBar';
+export { default as ScrollHighlightNav } from './NavBar/ScrollHighlightNav';
