+++ src/Containers/Header/Header.js
@@ -1,11 +1,9 @@
-import NavBar from '../../Components/NavBar/NavBar'
 import ScrollHighlightNav from '../../Components/NavBar/ScrollHighlightNav'
 import React from 'react';
 
 const Header = (props) => {
 
     return (
-        /*<NavBar />*/
         <ScrollHighlightNav links={props.links} />
     );
 }
