+++ src/Containers/index.js
@@ -4,3 +4,4 @@ export { default as Contact } from './Contact/Contact';
 export { default as Footer } from './Footer/Footer';
 export { default as Header } from './Header/Header';
 export { default as Home } from './Home/Home';
+export { default as Blog } from './Blog/Blog';
